src/requirements.txt
Django (2.0.4)
djangorestframework (3.8.0)
pip (9.0.1)
pytz (2018.3)
setuptools (28.8.0)
whitenoise (3.3.1)
